LayoutTests/http/tests/security/contentSecurityPolicy/allow-wasm-after-window-open-expected.txt

ALERT: PASS: wasm result is 7002
Test we allow web assembly with unsafe-eval after calling window.open.
